Still Confuse
That wasn’t where I put my site.

Maybe it was.

It was still returning ‘Hello’ so I’ve removed all of that PHP. But why
would I need to remove that!? It wouldn’t be there if I didn’t need it,
but it was overwriting all of my stuff? I’m so lost.